Move repository-related tests to dedicated file
