+++ include/codeowners/git_resources.hpp
@@ -41,17 +41,6 @@ discover_repository(const fs::path& location, bool cross_filesystems = false);
 repository_ptr
 create_repository(const fs::path& path);
 
-enum class discovery_strategy
-{
-    NEVER_CROSS_FILESYSTEMS,
-    CROSS_FILESYSTEMS
-};
-
-enum class repo_structure
-{
-    NONBARE,
-    BARE
-};
 
 /**
  * The attribute_set class holds a collection of file pattern-attribute value
