+++ src/repository.cpp
@@ -1,0 +1,5 @@
+#include <codeowners/repository.hpp>
+
+namespace co
+{
+} // end namespace 'co'
